New run

+++ scripts/21-hra-lit-construction.sh
@@ -1,9 +1,12 @@
 #!/bin/bash
 source constants.sh
 shopt -s extglob
-set -ev
 
-for f in hra-lit-kidney-as-nodes.csv hra-lit-kidney-disease-nodes.csv hra-lit-kidney-diseases-cooccur-edges.csv; do
-  rm -f input-csvs/$f
-  curl -s -o input-csvs/$f https://raw.githubusercontent.com/x-atlas-consortia/hra-lit/refs/heads/main/output-data/hra-lit/v0.7.1/reports/hra-lit-kidney/$f
-done
+if [ "$CLEAN" == "true" ]; then
+  set -ev
+
+  for f in hra-lit-kidney-as-nodes.csv hra-lit-kidney-disease-nodes.csv hra-lit-kidney-diseases-cooccur-edges.csv; do
+    rm -f input-csvs/$f
+    curl -s -o input-csvs/$f https://raw.githubusercontent.com/x-atlas-consortia/hra-lit/refs/heads/main/output-data/hra-lit/v0.7.1/reports/hra-lit-kidney/$f
+  done
+fi
